+++ core/Processors/KStreamProcessor.cs
@@ -1,11 +1,8 @@
-using Confluent.Kafka;
-using Microsoft.Extensions.Logging;
-using Streamiz.Kafka.Net.Crosscutting;
 using Streamiz.Kafka.Net.Processors.Public;
 
 namespace Streamiz.Kafka.Net.Processors
 {
-    internal class KStreamProcessor<K, V> : AbstractProcessor<K, V>
+    internal class KStreamProcessor<K, V> : KStreamPAPI<K, V>
     {
         private readonly Streamiz.Kafka.Net.Processors.Public.IProcessor<K, V> processor;
 
@@ -26,23 +23,7 @@ namespace Streamiz.Kafka.Net.Processors
             processor.Close();
         }
 
-        public override void Process(K key, V value)
-        {
-            if (key == null && StateStores.Count > 0)
-            {
-                log.LogWarning($"Skipping record due to null key because your processor is stateful. topic=[{Context.Topic}] partition=[{Context.Partition}] offset=[{Context.Offset}]");
-                droppedRecordsSensor.Record();
-                return;
-            }
-            
-            Record<K, V> record = new Record<K, V>(
-                new TopicPartitionOffset(Context.Topic, Context.Partition, Context.Offset),
-                Context.RecordContext.Headers,
-                new Timestamp(Context.RecordContext.Timestamp.FromMilliseconds()),
-                key,
-                value);
-            
-            processor.Process(record);
-        }
+        public override void Process(Record<K, V> record)
+            => processor.Process(record);
     }
 }
